ミグレイトファイル安全化②

+++ db/migrate/20251217074509_remove_legacy_columns_from_reflections.rb
@@ -1,8 +1,5 @@
 class RemoveLegacyColumnsFromReflections < ActiveRecord::Migration[7.1]
   def change
-    remove_column :reflections, :date, :date
-    remove_column :reflections, :today_task, :text
-    remove_column :reflections, :review1, :text
-    remove_column :reflections, :review2, :text
+    remove_column :reflections, :date, :date if column_exists?(:reflections, :date)
   end
 end
